MinimalExample

This class is a very basic example of the features of the rJava maven plugin. <br>

Version: 0.02

Generated: 2021-12-23T00:04:10.395

Details

The class is annotated with an @RClass to identify it as part of the R API. <br>

Methods

Constructors

J$MinimalExample$new()

Static methods

none

Instance methods

instance$demo(dataframe, message)

instance$clone()

instance$print()

Method new()

the default no-args constructor

Usage

J = testRapi::JavaApi$get()
J$MinimalExample$new()

Arguments

none

Returns

R6 MinimalExample object:

Examples

not available

Method demo()

Documentation of the method can be done in JavaDoc and these will be present in the R documentation

Usage

J = testRapi::JavaApi$get()
instance = J$MinimalExample$new(...);
instance$demo(dataframe, message)

Arguments

dataframe - a dataframe with an arbitrary number of columns - (java expects a RDataframe)

message - a message - (java expects a String)

Returns

RDataframe: the dataframe unchanged

Examples

minExample = J$MinimalExample$new()
minExample$demo(dataframe=tibble::tibble(input=c(1,2,3)), message='Hello world')

Examples

if (FALSE) {
J = testRapi::JavaApi$get();
instance = J$MinimalExample$new()

## -----------------------------------
## Method `MinimalExample$demo`
## -----------------------------------
  
    minExample = J$MinimalExample$new()
minExample$demo(dataframe=tibble::tibble(input=c(1,2,3)), message='Hello world')

}