Serialiser

Actually useful class which takes an R dataframe and a filename

Version: 0.02

Generated: 2021-12-27T18:17:36.641

Details

string and serialises the dataframe so it can be used natively in java for testing purposes.

Methods

Constructors

J$Serialiser$new()

Static methods

J$Serialiser$serialiseDataframe(dataframe, filename)

J$Serialiser$deserialiseDataframe(filename)

J$Serialiser$serialiseList(dataframe, filename)

J$Serialiser$deserialiseList(filename)

J$Serialiser$serialiseNamedList(dataframe, filename)

J$Serialiser$deserialiseNamedList(filename)

Instance methods

none

instance$clone()

instance$print()

Method new()

the default no-args constructor

Usage

J = testRapi::JavaApi$get()
J$Serialiser$new()

Arguments

none

Returns

R6 Serialiser object:

Examples

not available

Method serialiseDataframe()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$serialiseDataframe(dataframe, filename)

Arguments

dataframe - (java expects a RDataframe)

filename - (java expects a String)

Returns

void:

Examples

not available

Method deserialiseDataframe()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$deserialiseDataframe(filename)

Arguments

filename - (java expects a String)

Returns

RDataframe:

Examples

not available

Method serialiseList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$serialiseList(dataframe, filename)

Arguments

dataframe - (java expects a RList)

filename - (java expects a String)

Returns

void:

Examples

not available

Method deserialiseList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$deserialiseList(filename)

Arguments

filename - (java expects a String)

Returns

RList:

Examples

not available

Method serialiseNamedList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$serialiseNamedList(dataframe, filename)

Arguments

dataframe - (java expects a RNamedList)

filename - (java expects a String)

Returns

void:

Examples

not available

Method deserialiseNamedList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$deserialiseNamedList(filename)

Arguments

filename - (java expects a String)

Returns

RNamedList:

Examples

not available

Examples

if (FALSE) {
J = testRapi::JavaApi$get();
instance = J$Serialiser$new()

## -----------------------------------
## Method `Serialiser$serialiseDataframe`
## -----------------------------------

## -----------------------------------
## Method `Serialiser$deserialiseDataframe`
## -----------------------------------

## -----------------------------------
## Method `Serialiser$serialiseList`
## -----------------------------------

## -----------------------------------
## Method `Serialiser$deserialiseList`
## -----------------------------------

## -----------------------------------
## Method `Serialiser$serialiseNamedList`
## -----------------------------------

## -----------------------------------
## Method `Serialiser$deserialiseNamedList`
## -----------------------------------

}