FactoryTest

missing description

Version: 0.2.0.9000

Generated: 2022-05-23T15:59:29.667162

Details

no details

Methods

Constructors

J$FactoryTest$new()

Static methods

none

Instance methods

instance$generateCharacter()

instance$generateNumeric()

instance$generateInteger()

instance$generateFactor()

instance$generateLogical()

instance$generateCharacterVec()

instance$generateNumericVec()

instance$generateIntegerVec()

instance$generateFactorVec()

instance$generateLogicalVec()

instance$generateDataframe()

instance$generateStreamDataframe()

instance$generateList()

instance$generateNamedList()

instance$clone()

instance$print()

Method new()

the default no-args constructor

Usage

J = testRapi::JavaApi$get()
J$FactoryTest$new()
        

Arguments

none

Returns

R6 FactoryTest object:

Examples

not available

Method generateCharacter()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateCharacter()
        

Arguments

none

Returns

RCharacter:

Examples

not available

Method generateNumeric()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateNumeric()
        

Arguments

none

Returns

RNumeric:

Examples

not available

Method generateInteger()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateInteger()
        

Arguments

none

Returns

RInteger:

Examples

not available

Method generateFactor()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateFactor()
        

Arguments

none

Returns

RFactor:

Examples

not available

Method generateLogical()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateLogical()
        

Arguments

none

Returns

RLogical:

Examples

not available

Method generateCharacterVec()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateCharacterVec()
        

Arguments

none

Returns

RCharacterVector:

Examples

not available

Method generateNumericVec()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateNumericVec()
        

Arguments

none

Returns

RNumericVector:

Examples

not available

Method generateIntegerVec()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateIntegerVec()
        

Arguments

none

Returns

RIntegerVector:

Examples

not available

Method generateFactorVec()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateFactorVec()
        

Arguments

none

Returns

RFactorVector:

Examples

not available

Method generateLogicalVec()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateLogicalVec()
        

Arguments

none

Returns

RLogicalVector:

Examples

not available

Method generateDataframe()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateDataframe()
        

Arguments

none

Returns

RDataframe:

Examples

not available

Method generateStreamDataframe()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateStreamDataframe()
        

Arguments

none

Returns

RDataframe:

Examples

not available

Method generateList()

Lists are much harder to type check than vectors hence RList builder methods throw checked exceptions

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateList()
        

Arguments

none

Returns

RList: a RList containing the supplied Java objects converted into RObjects

Examples

not available

Method generateNamedList()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateNamedList()
        

Arguments

none

Returns

RNamedList:

Examples

not available

Examples

if (FALSE) {
J = testRapi::JavaApi$get();
instance = J$FactoryTest$new()

## -----------------------------------
## Method `FactoryTest$generateCharacter`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNumeric`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateInteger`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateFactor`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateLogical`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateCharacterVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNumericVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateIntegerVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateFactorVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateLogicalVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateDataframe`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateStreamDataframe`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateList`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNamedList`
## -----------------------------------

}