FeatureTest

A test of the R6 generator templating

Version: 0.2.0.9000

Generated: 2022-05-23T16:14:42.849396

Arguments

logMessage
logMessage - a message which will be logged - (java expects a String)

Details

The feature test should allow mathjax in javadoc

$$e = mc^2$$

this is a details comment

Methods

Constructors

J$FeatureTest$new(logMessage)

Static methods

J$FeatureTest$demoStatic(message)

J$FeatureTest$diamonds()

Instance methods

instance$doHelloWorld()

instance$doSum(a, b)

instance$doSum2(a, b)

instance$getMessage()

instance$fluentSetMessage(message)

instance$factoryMethod(a, b)

instance$objectAsParameter(otherObj)

instance$doSomethingWithDataFrame(dataframe)

instance$generateDataFrame()

instance$errorThrower()

instance$clone()

instance$print()

Method new()

A maximum of one constructor of any signature can be used. <br>

If different constructors are needed then they may be used but not included in the R Api (i.e. not annotated with @RMethod.) <br>

Static factory methods can be used instead.

Usage

J = testRapi::JavaApi$get()
J$FeatureTest$new(logMessage)
        

Arguments

logMessage - a message which will be logged - (java expects a String)

Returns

R6 FeatureTest object:

Examples

  
      minExample = J$FeatureTest$new('Hello from Java constructor!')
      

Method doHelloWorld()

Description of a hello world function

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$doHelloWorld()
        

Arguments

none

Returns

RCharacter: this java method returns a String

Examples

minExample = J$FeatureTest$new('Hello, R World!')
minExample$doHelloWorld()
      

Method doSum()

The doSum function description = it adds two numerics

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$doSum(a, b)
        

Arguments

a the A parameter, can be NA - (java expects a RNumeric)

b the B parameter - (java expects a RNumeric)

Returns

RNumeric: A+B of course, NAs in inputs are converted to null in Java. This catches the resulting NPE in java idiom and returns an explicit NA. This only matters if you care about the difference between NA_real_ and NaN in R.

Examples

not available

Method doSum2()

Do sum 2 uses native ints rather than RNumerics It should throw an error if given something that cannot be coerced to an integer. This also demonstrated the use of the `@RDefault` annotation

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$doSum2(a, b)
        

Arguments

a the A parameter - (java expects a int)

b the B parameter - (defaulting to 10) - (java expects a int)

Returns

int: A+B of course

Examples

not available

Method demoStatic()

Static methods are also supported. These are accessed through the root of the R api.

Usage

J = testRapi::JavaApi$get()
J$FeatureTest$demoStatic(message)
        

Arguments

message a message - (java expects a String)

Returns

void:

Examples

J$FeatureTest$demoStatic('Ola, el mundo')
      

Method getMessage()

message desciption

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$getMessage()
        

Arguments

none

Returns

RCharacter: The message (previously set by the constructor)

Examples

not available

Method fluentSetMessage()

A fluent method which updates the message in this object, returning the same object. This is differentiated from factory methods which produce a new instance of the same class by checking to see if the returned Java object is equal to the calling Java object.

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$fluentSetMessage(message)
        

Arguments

message the message is a string - (defaulting to \"hello\nworld\") - (java expects a RCharacter)

Returns

R6 FeatureTest object: this should return exactly the same R6 object.

Examples

not available

Method factoryMethod()

A factory or builder method which constructs an object of another class from some parameters

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$factoryMethod(a, b)
        

Arguments

a the first parameter - (java expects a RCharacter)

b the second parameter - (defaulting to as.character(Sys.Date())) - (java expects a RCharacter)

Returns

R6 MoreFeatureTest object: A MoreFeatureTest R6 reference

Examples

not available

Method objectAsParameter()

no description

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$objectAsParameter(otherObj)
        

Arguments

otherObj - (java expects a MoreFeatureTest)

Returns

String:

Examples

not available

Method doSomethingWithDataFrame()

Consumes a data frame and logs its length

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$doSomethingWithDataFrame(dataframe)
        

Arguments

dataframe a dataframe - (java expects a RDataframe)

Returns

void:

Examples

not available

Method generateDataFrame()

Creates a basic dataframe and returns it

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$generateDataFrame()
        

Arguments

none

Returns

RDataframe: a daatframe

Examples

not available

Method diamonds()

A copy of the ggplot2::diamonds dataframe serialised into java, using RObject.writeRDS, saved within the jar file of the package, and exposed here using RObject.readRDS.

Usage

J = testRapi::JavaApi$get()
J$FeatureTest$diamonds()
        

Arguments

none

Returns

RDataframe: the ggplot2::diamonds dataframe

Examples

J$FeatureTest$diamonds()
      

Method errorThrower()

just throws an error

Usage

J = testRapi::JavaApi$get()
instance = J$FeatureTest$new(...);
instance$errorThrower()
        

Arguments

none

Returns

void:

Examples

not available

Examples

if (FALSE) {
J = testRapi::JavaApi$get();
instance = J$FeatureTest$new(logMessage)

## -----------------------------------
## Method `FeatureTest$doHelloWorld`
## -----------------------------------
minExample = J$FeatureTest$new('Hello, R World!')
minExample$doHelloWorld()

## -----------------------------------
## Method `FeatureTest$doSum`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$doSum2`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$demoStatic`
## -----------------------------------
J$FeatureTest$demoStatic('Ola, el mundo')

## -----------------------------------
## Method `FeatureTest$getMessage`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$fluentSetMessage`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$factoryMethod`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$objectAsParameter`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$doSomethingWithDataFrame`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$generateDataFrame`
## -----------------------------------

## -----------------------------------
## Method `FeatureTest$diamonds`
## -----------------------------------
J$FeatureTest$diamonds()

## -----------------------------------
## Method `FeatureTest$errorThrower`
## -----------------------------------

}