A test library

This package is a feature test of the R6 Generator Maven Plugin. It is not intended to be useful beyond testing, demonstrating and documenting the features of the plugin.

Version: 0.2.0.9000

Generated: 2022-05-23T17:14:54.487780

Arguments

logLevel
optional - the slf4j log level as a string - one of OFF (most specific, no logging), FATAL (most specific, little data), ERROR, WARN, INFO, DEBUG, TRACE (least specific, a lot of data), ALL (least specific, all data)

Usage

 J = testRapi::JavaApi$get(logLevel)
   

Author

rc538@exeter.ac.uk

Static methods and constructors

JavaApi$get()

J$MoreFeatureTest$new(message1, message2)

J$MoreFeatureTest$create(message1, message2)

J$MoreFeatureTest$concat(message1, message2)

J$BounceTest$new()

J$FeatureTest$new(logMessage)

J$FeatureTest$demoStatic(message)

J$FeatureTest$diamonds()

J$FactoryTest$new()

J$Serialiser$new()

J$Serialiser$serialiseDataframe(dataframe, filename)

J$Serialiser$deserialiseDataframe(filename)

J$Serialiser$serialiseList(dataframe, filename)

J$Serialiser$deserialiseList(filename)

J$Serialiser$serialiseNamedList(dataframe, filename)

J$Serialiser$deserialiseNamedList(filename)

J$MinimalExample$new()

Method MoreFeatureTest$new()

the first constructor is used if there are none annotated

Usage

J = testRapi::JavaApi$get()
J$MoreFeatureTest$new(message1, message2)
        

Arguments

message1 - the message to be printed - (java expects a RCharacter) message2 - will be used for toString - (java expects a RCharacter)

Returns

R6 MoreFeatureTest object:

Method MoreFeatureTest$create()

A static object constructor

Usage

J = testRapi::JavaApi$get()
J$MoreFeatureTest$create(message1, message2)
        

Arguments

message1 - the message to be printed - (java expects a RCharacter) message2 - will be used for toString - (java expects a RCharacter)

Returns

R6 MoreFeatureTest object: A MoreFeatureTest R6 object

Method MoreFeatureTest$concat()

no description

Usage

J = testRapi::JavaApi$get()
J$MoreFeatureTest$concat(message1, message2)
        

Arguments

message1 - (java expects a RCharacter) message2 - (java expects a RCharacter)

Returns

RCharacter:

Method BounceTest$new()

the default no-args constructor

Usage

J = testRapi::JavaApi$get()
J$BounceTest$new()
        

Arguments

none

Returns

R6 BounceTest object:

Method FeatureTest$new()

A maximum of one constructor of any signature can be used. <br>

If different constructors are needed then they may be used but not included in the R Api (i.e. not annotated with @RMethod.) <br>

Static factory methods can be used instead.

Usage

J = testRapi::JavaApi$get()
J$FeatureTest$new(logMessage)
        

Arguments

logMessage - a message which will be logged - (java expects a String)

Returns

R6 FeatureTest object:

Method FeatureTest$demoStatic()

Static methods are also supported. These are accessed through the root of the R api.

Usage

J = testRapi::JavaApi$get()
J$FeatureTest$demoStatic(message)
        

Arguments

message a message - (java expects a String)

Returns

void:

Method FeatureTest$diamonds()

A copy of the ggplot2::diamonds dataframe serialised into java, using RObject.writeRDS, saved within the jar file of the package, and exposed here using RObject.readRDS.

Usage

J = testRapi::JavaApi$get()
J$FeatureTest$diamonds()
        

Arguments

none

Returns

RDataframe: the ggplot2::diamonds dataframe

Method FactoryTest$new()

the default no-args constructor

Usage

J = testRapi::JavaApi$get()
J$FactoryTest$new()
        

Arguments

none

Returns

R6 FactoryTest object:

Method Serialiser$new()

the default no-args constructor

Usage

J = testRapi::JavaApi$get()
J$Serialiser$new()
        

Arguments

none

Returns

R6 Serialiser object:

Method Serialiser$serialiseDataframe()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$serialiseDataframe(dataframe, filename)
        

Arguments

dataframe - (java expects a RDataframe) filename - (java expects a String)

Returns

void:

Method Serialiser$deserialiseDataframe()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$deserialiseDataframe(filename)
        

Arguments

filename - (java expects a String)

Returns

RDataframe:

Method Serialiser$serialiseList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$serialiseList(dataframe, filename)
        

Arguments

dataframe - (java expects a RList) filename - (java expects a String)

Returns

void:

Method Serialiser$deserialiseList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$deserialiseList(filename)
        

Arguments

filename - (java expects a String)

Returns

RList:

Method Serialiser$serialiseNamedList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$serialiseNamedList(dataframe, filename)
        

Arguments

dataframe - (java expects a RNamedList) filename - (java expects a String)

Returns

void:

Method Serialiser$deserialiseNamedList()

no description

Usage

J = testRapi::JavaApi$get()
J$Serialiser$deserialiseNamedList(filename)
        

Arguments

filename - (java expects a String)

Returns

RNamedList:

Method MinimalExample$new()

the default no-args constructor

Usage

J = testRapi::JavaApi$get()
J$MinimalExample$new()
        

Arguments

none

Returns

R6 MinimalExample object: